Extract sidebar component + cleanup
